+++ src/components/RichTextEditor/serializer/markdown/deserializer/index.ts
@@ -1,0 +1,11 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+/**
+ * The main source
+ * https://github.com/hanford/remark-slate
+ */
+import deserialize from '../deserializer/deserialize';
+import plugin from './plugin';
+
+export { deserialize };
+export default plugin;
