+++ src/components/RichTextEditor/serializer/markdown/serializer/index.ts
@@ -4,10 +4,6 @@
  * The main source
  * https://github.com/hanford/remark-slate
  */
-import deserialize from './deserialize';
 import serialize from './serialize';
-import plugin from './plugin';
 
-export * from './astTypes';
-export { deserialize, serialize };
-export default plugin;
+export { serialize };
