+++ src/components/RichTextEditor/serializer/markdown/serializer/isLeafNode.ts
@@ -1,5 +1,5 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { LeafType, NodeType } from '../astTypes';
+import { LeafType, NodeType } from '../types';
 
 export const isLeafNode = (node: NodeType): node is LeafType => !('type' in node);
